Add 'yggdrasil' to ./configure script(s)
The easiest way to compile is to run "./configure" and have it
generate all the required "Makefile" files. For some reason,
nobody thought of having ./configure detect whatever protocols
are included under "libpurple/protocols" and dynamically and
the pidgin project continues using the old hardcoded method.

I suppose it works well enough for whatever their development
process happens to be, but, it's absolutely terrible for future
growth when it comes to adding new protocols because it's simply
a painful task to have to find-and-append all the new protocols.